requirements.txt
pandas
streamlit
numpy
BytesIO
socket
plotly.express
re
selenium.webdriver
time
sqlalchemy